2nd part of changes due to comments for PR
