+++ TODO.md
@@ -6,4 +6,5 @@
 - cat-split fusion
 - working on all tests
 - Adding annotation library installation and removing ad-hoc import of the latter
+- clean up utils for annotations
 - Changing PaSh flags (making the default be priority r-split and then consecutive chunks), so remove the r_split flag and make defaults be the ones from the OSDI paper (I can do that TODO too)
