+++ compiler/annotations_utils/util_file_descriptors.py
@@ -1,14 +1,8 @@
 from util import log
-from definitions.ir.file_id import FileId
 from definitions.ir.resource import FileResource, Resource, FileDescriptorResource
-from definitions.ir.arg import Arg
-# imports from annotation framework
 import sys
 from config import get_path_annotation_repo
 sys.path.insert(1, get_path_annotation_repo())
-# for typing
-# --
-# for use
 from datatypes_new.BasicDatatypesWithIO import FileNameWithIOInfo, StdDescriptorWithIOInfo
