+++ compiler/annotations_utils/util_mapper.py
@@ -1,3 +1,5 @@
+# TODO: this file can properly be deleted
+
 # imports from annotation framework
 import sys
 from config import get_path_annotation_repo
@@ -7,14 +9,16 @@ sys.path.insert(1, get_path_annotation_repo())
 from annotation_generation_new.datatypes.parallelizability.Mapper import Mapper
 
 from definitions.ir.dfg_node import DFGNode
-from util_new_cmd_invocations import get_command_invocation_prefix_from_dfg_node
+from annotations_utils.util_cmd_invocations import get_command_invocation_prefix_from_dfg_node
 from util import log
 
 def get_actual_mapper_from_node(node, parallelizer) -> Mapper:
+    assert(False)
     cmd_inv_pref = get_command_invocation_prefix_from_dfg_node(node)
     return parallelizer.get_actual_mapper(cmd_inv_pref)
 
 def get_mapper_as_dfg_node_from_node(node, parallelizer, inputs, outputs) -> DFGNode:
+    assert(False)
     mapper = get_actual_mapper_from_node(node, parallelizer)
     log(f'mapper for cmd_name: {node.com_name}')
     log(f'here mapper: {mapper}')
@@ -46,6 +50,7 @@ def get_mapper_as_dfg_node_from_node(node, parallelizer, inputs, outputs) -> DFG
 # def get_map_output_files(node, input_edge_ids, fileIdGen):
 # NEW
 def get_map_output_files(node, input_edge_ids, fileIdGen, parallelizer):
+    assert(False)
     assert (node.is_parallelizable())
     # TODO ANNO: How to substitute? @KK
     if (node.com_category == "stateless"):
@@ -70,6 +75,7 @@ def get_map_output_files(node, input_edge_ids, fileIdGen, parallelizer):
 # def pure_get_map_output_files(node, input_edge_ids, fileIdGen):
 # NEW
 def pure_get_map_output_files(node, input_edge_ids, fileIdGen, parallelizer):
+    assert(False)
     assert (node.is_pure_parallelizable())
     # BEGIN ANNO
     # OLD
